// Backend/infrastructure/persistence/mongodb/schemas/subscribtion.ts
import { ObjectId } from 'mongodb';

export interface ISubscribtion {
    _id?: ObjectId;
    name: string
    price: number;
    duration: 'Monthly' | 'Quarterly' | 'Annual';
    isActive: boolean;
    interviewLimit: number;
    interviewUnlimited: boolean;
    jobPostLimit: number;
    jobUnlimited: boolean;
    hasAI: boolean;
    createdAt: Date;
    updatedAt: Date;
}